archive/early_prototype_snapshot/requirements.txt
appnope==0.1.4
asttokens==3.0.1
comm==0.2.3
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.8.0
jupyter_core==5.9.1
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
packaging==26.0
parso==0.8.6
pexpect==4.9.0
platformdirs==4.9.4
prompt_toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyarrow
python-dateutil==2.9.0.post0
pyzmq==27.1.0
qiskit
qiskit-algorithms
qiskit-machine-learning
six==1.17.0
stack-data==0.6.3
tornado==6.5.4
traitlets==5.14.3
wcwidth==0.6.0
